refactor(wip): Re-organize API code.
API related scripts moved to scripts/api_base.py

File 'scripts/user_api/api.py' only contains API classes, methods and
attributes visible/accessible to users.
