+++ py2hwsw/scripts/user_api/api.py
@@ -17,6 +17,7 @@
 # may call special conversion methods instead of setting the attribute directly.
 
 
+# Note for developers:
 # To achieve backwards compatibility, consider the following best practices:
 #
 # - Add new methods, don't change existing ones:
@@ -35,55 +36,31 @@
 #       understand the impact of updates and plan accordingly.
 
 
+# TODO: Remove these
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from pydantic import BaseModel
+
 from datetime import date
 import os
 import sys
-import inspect
+
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(os.path.realpath(__file__)), '..')))
+from api_base import api_for, empty_list, empty_dict
 import iob_conf as internal_conf
 
-#
-# Utilities used by API
-#
-# TODO: Move these utilities to another file (maybe iob_base.py)
 
+# NOTE: for developers:
+# - Update py2hwsw version every time API changes!
+#   Must change major version if new API is not backwards compatible.
+#   Change minor version when adding new API methods or other non-breaking changes.
 
-def _empty_list():
-    return field(default_factory=list)
-
-
-def _empty_dict():
-    return field(default_factory=dict)
-
-
-class ValidatingBaseModel(BaseModel):
-    """
-    Similar to BaseModel class of pydantic, but with validate_assignment=True.
-    Also add '__post_init__' method for subclasses, for similar purpose as dataclasses __post_init__.
-    """
-
-    # Configure validate_assignment to validate all attributes when they are set.
-    class Config:
-        validate_assignment = True
-
-    def __init__(self, **data):
-        super().__init__(**data)
-        # Additional initialization logic here
-        self.__post_init__(**data)
-
-    def __post_init__(self, **data):
-        """
-        A __post_init__ method, similar to the one in dataclasses, that can be overridden in subclasses.
-        """
-        pass
-
-
-# NOTE: Update py2hwsw version every time API changes!
-#       Must change major version if new API is not backwards compatible.
-#       Change minor version when adding new API methods or other non-breaking changes.
+# NOTE: for users:
+# - Every attribute must be accessed via setter/getter methods.
+#   Users should use corresponding getters/setters [like `get_<attribute_name>()`/`set_<attribute_name>()`].
+#   Some of these methods may be automatically generated by the decorator.
+# - Classes have similar structure to python's '@dataclasses'
+# - Methods are abstract. They do not contain implementation in the API.
 
 #
 # Confs
@@ -110,177 +87,6 @@ conf_short2python = [
     ["-m", "min_value"],
     ["-M", "max_value"],
 ]
-
-
-# These functions should be called by py2 every time it exchanges objects with user.
-def convert2api(internal_obj):
-    """
-    Convert given internal object to corresponding API object.
-
-    Attributes:
-        internal_obj (object): internal object
-    Returns:
-        object: api object
-    """
-    pass
-
-
-def convert2internal(api_obj):
-    """
-    Convert given API object to corresponding internal object.
-
-    Attributes:
-        api_obj (object): API object
-    Returns:
-        object: internal object
-    """
-    pass
-
-
-def get_methods(cls):
-    """
-    Returns all non-default methods in a class (includes ones inherited from superclasses)
-
-    Attributes:
-        cls (class): class
-    Returns:
-        dict: methods of the class and their type hints
-    """
-    methods = {}
-    for name in dir(cls):
-        attr = getattr(cls, name)
-        if inspect.isfunction(attr) and not name.startswith("__"):
-            signature = inspect.signature(attr)
-            methods[name] = signature
-    return methods
-
-
-def has_body(func):
-    """Check if a function has a body."""
-    source = inspect.getsource(func)
-    # Remove leading and trailing whitespace
-    source = source.strip()
-    # Remove the function definition line
-    lines = source.split('\n')
-    if len(lines) < 2:
-        return False
-    body = '\n'.join(lines[1:])
-    return body.strip() != 'pass'
-
-
-def api_for(internal_cls):
-    """
-    Decorator for creating API interface. Apply this decorator to every function in the API.
-
-    This decorator:
-    1) Passes attributes defined in the API class to the internal class (via arguments to the internal classes's constructor)
-    2) Creates setters/getters for every attribute of API class, but actually accessing internal classes's attributes.
-    3) Replaces (abstract) API methods with functions that call the corresponding internal class methods.
-    4) Removes attributes from API class. This prevents user from trying to access attributes directly.
-
-    From point of view of the user (even when checking api.py code), he only sees API class with the (getters/setters of the) attributes and methods defined in it.
-    The user may instantiate any API class and use its attributes/methods. This decorator will handle the instantiation of the internal class and hide it from user.
-
-    With this decorator, when the user instantiates API class, two objects are created: the API object and the internal object.
-    The API object is returned to user. The internal object is stored internally by py2.
-    The API and internal objects have a unique id and are associated with each other in a mapping stored internally by py2.
-
-    When py2 needs to pass objects to the user, py2 should only pass API objects (converting from internal objects if needed).
-    When the user needs to pass objects to py2, he passes API objects, which py2 will convert to internal objects.
-
-    Attributes:
-        internal_cls (class): Reference to internal class that will extend functionality of API class being decorated.
-    Returns:
-        class: Updated API class
-    """
-
-    # Create decorator dynamically for the API class
-    def decorator(cls):
-        # Get attributes and their default values
-        attributes = {k: cls.__dict__[k] for k in cls.__annotations__}
-        methods = get_methods(cls)
-
-        # Update constructor of the API class
-        def new_init(self, *args, **kwargs):
-            print("API class constructor called: ", cls.__name__)
-            print("Attributes: ", attributes) # Dictionary with attributes and their types
-            print("Annotations: ", cls.__annotations__) # Dictionary with attributes and their types
-            print("Methods: ", methods) # Dictionary with methods and their types
-
-            # Instantiate internal class with API attributes
-            internal_obj = internal_cls(attributes, cls.__annotations__, args, kwargs)
-
-            # Store reference to internal object
-            self.__internal_obj = internal_obj
-
-        cls.__init__ = new_init
-
-        # Local utility functions
-        def _generate_setter(attribute_name):
-            """
-            Function to generate setter for a given attribute.
-            Setter will pass the given value to the corresponding attribute in the internal object.
-            """
-
-            def setter(self, value):
-                # print("Setter called for attribute: ", attribute_name)
-                setattr(self.__internal_obj, attribute_name, value)
-            return setter
-
-        def _generate_getter(attribute_name):
-            """
-            Function to generate getter for a given attribute.
-            Getter will get value from the corresponding attribute in the internal object.
-            """
-
-            def getter(self):
-                # print("Getter called for attribute: ", attribute_name)
-                return getattr(self.__internal_obj, attribute_name)
-            return getter
-
-        def _generate_method_wrapper(method_name):
-            """
-            Function to generate wrapper for a given method.
-            Wrapper will call the corresponding method in the internal object.
-            """
-
-            def wrapper(self, *args, **kwargs):
-                assert hasattr(self.__internal_obj, method_name), f"[Py2HWSW bug]: Missing implementation for API method '{method_name}' of class '{cls.__name__}'!"
-                # print("Method called: ", method_name)
-                return getattr(self.__internal_obj, method_name)(*args, **kwargs)
-            return wrapper
-
-        # Generate and add setters/getters for each attribute of API class, referencing values from corresponding attribute in internal object
-        for name in cls.__annotations__:
-            setattr(cls, f"set_{name}", _generate_setter(name))
-            setattr(cls, f"get_{name}", _generate_getter(name))
-
-        # Replace API methods by calls to the internal methods
-        for name in methods:
-            # Ensure method is abstract (does not have body) in API class
-            assert not has_body(getattr(cls, name)), f"API method '{name}' must be abstract."
-
-            # Call corresponding method in internal class
-            setattr(cls, name, _generate_method_wrapper(name))
-
-        # Remove class attributes from API class
-        # This will make sure user cannot access them directly.
-        for name in cls.__annotations__:
-            delattr(cls, name)
-
-        return cls
-
-    return decorator
-
-#
-# API classes, attributes, and methods
-#
-# Notes:
-# - Every attribute must be accessed via setter/getter methods.
-#   Users should use corresponding getters/setters [like `get_<attribute_name>()`/`set_<attribute_name>()`].
-#   Some of these methods may be automatically generated by the decorator.
-# - Classes have similar structure to python's '@dataclasses'
-# - Methods are abstract. They do not contain implementation.
 
 
 @api_for(internal_conf.iob_conf)
@@ -346,7 +152,7 @@ class iob_conf_group(ABC):
 
     name: str = ""
     descr: str = "Default description"
-    confs: list[iob_conf] = _empty_list()
+    confs: list[iob_conf] = empty_list()
     doc_only: bool = False
     doc_clearpage: bool = False
 
@@ -395,7 +201,7 @@ class iob_signal(ABC):
     # # Used for `iob_comb`: If enabled, iob_comb will infer a register for this signal.
     # isreg: bool = False
     # # Used for `iob_comb`: List of signals associated to the infered register.
-    # reg_signals: list[str] = _empty_list()
+    # reg_signals: list[str] = empty_list()
 
     # Logic value for future simulation effort using global signals list.
     # See 'TODO' in iob_core.py for more info: https://github.com/IObundle/py2hwsw/blob/a1e2e2ee12ca6e6ad81cc2f8f0f1c1d585aaee73/py2hwsw/scripts/iob_core.py#L251-L259
@@ -443,7 +249,7 @@ class interface(ABC):
     prefix: str = ""
     mult: str | int = 1
     params: str = None
-    widths: dict[str, str] = _empty_dict()
+    widths: dict[str, str] = empty_dict()
     file_prefix: str = ""
     portmap_port_prefix: str = ""
 
@@ -490,7 +296,7 @@ class iob_wire(ABC):
     descr: str = "Default description"
     if_defined: str = ""
     if_not_defined: str = ""
-    signals: list[iob_signal] = _empty_list()
+    signals: list[iob_signal] = empty_list()
 
     @abstractmethod
     def get_signal(self, signal_name: str) -> iob_signal:
@@ -539,7 +345,7 @@ class iob_port(iob_wire):
 
     e_connect: iob_wire = None
     # FIXME: Constrain values received by bit slices
-    e_connect_bit_slices: list[object] = _empty_list()
+    e_connect_bit_slices: list[object] = empty_list()
     doc_only: bool = False
     doc_clearpage: bool = False
 
@@ -672,7 +478,7 @@ class iob_block_group(ABC):
 
     name: str = ""
     descr: str = "Default description"
-    blocks: list[iob_core] = _empty_list()
+    blocks: list[iob_core] = empty_list()
     doc_clearpage: bool = False
 
 
@@ -762,15 +568,15 @@ class iob_module(ABC):
     name: str = ""
     description: str = "Default description"
     reset_polarity: str = None
-    confs: list[iob_conf_group] = _empty_list()
-    ports: list[iob_port] = _empty_list()
-    wires: list[iob_wire] = _empty_list()
-    snippets: list[iob_snippet] = _empty_list()
+    confs: list[iob_conf_group] = empty_list()
+    ports: list[iob_port] = empty_list()
+    wires: list[iob_wire] = empty_list()
+    snippets: list[iob_snippet] = empty_list()
     comb: iob_comb = None
     fsm: iob_fsm = None
-    subblocks: list[iob_block_group] = _empty_list()
-    superblocks: list[iob_block_group] = _empty_list()
-    sw_modules: list[iob_block_group] = _empty_list()
+    subblocks: list[iob_block_group] = empty_list()
+    superblocks: list[iob_block_group] = empty_list()
+    sw_modules: list[iob_block_group] = empty_list()
 
 
 # Convert dict keys to python attributes
@@ -811,11 +617,11 @@ class iob_instance(ABC):
 
     instance_name: str = None
     instance_description: str = ""
-    parameters: dict[str, int | str] = _empty_dict()
+    parameters: dict[str, int | str] = empty_dict()
     if_defined: str = ""
     if_not_defined: str = ""
     instantiate: bool = ""
-    connect: dict[str, str] = _empty_dict()
+    connect: dict[str, str] = empty_dict()
 
 
 # Convert dict keys to python attributes
@@ -899,15 +705,15 @@ class iob_core(iob_module, iob_instance):
     # instance_name: str = None
     use_netlist: bool = False
     is_system: bool = False
-    board_list: list[str] = _empty_list()
+    board_list: list[str] = empty_list()
     # dest_dir: str = "" # Should not be visible to the user
-    ignore_snippets: list[str] = _empty_list()
+    ignore_snippets: list[str] = empty_list()
     generate_hw: bool = False
-    parent: dict = _empty_dict()  # FIXME: not sure if this will be needed in pythase?
+    parent: dict = empty_dict()  # FIXME: not sure if this will be needed in pythase?
     is_top_module: bool = False
     is_superblock: bool = False
     is_tester: bool = False
-    python_parameters: list[object] = _empty_list()
+    python_parameters: list[object] = empty_list()
     license: iob_license = None
     doc_conf: str = ""
     title: str = ""
